add maven, fix code error
