+++ src/main/java/com/geetest/sdk/java/msg/demo/VerifyGeetestServlet.java
@@ -1,15 +1,14 @@
 package com.geetest.sdk.java.msg.demo;
 
-import java.io.IOException;
-import java.io.PrintWriter;
+import com.geetest.sdk.java.GeetestLib;
+import com.geetest.sdk.java.GeetestMsgLib;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
-import com.geetest.sdk.java.GeetestLib;
-import com.geetest.sdk.java.GeetestMsgLib;
+import java.io.IOException;
+import java.io.PrintWriter;
 
 /**
  * Msg套件下的拼图一次验证
