+++ src/main/java/com/geetest/sdk/java/msg/demo/VerifyMsgServlet.java
@@ -1,14 +1,13 @@
 package com.geetest.sdk.java.msg.demo;
 
-import java.io.IOException;
-import java.io.PrintWriter;
+import com.geetest.sdk.java.GeetestMsgLib;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
-import com.geetest.sdk.java.GeetestMsgLib;
+import java.io.IOException;
+import java.io.PrintWriter;
 
 /**
  * 二次的短信验证
