+++ src/main/java/com/geetest/sdk/java/web/demo/GeetestConfig.java
@@ -9,9 +9,12 @@ package com.geetest.sdk.java.web.demo;
 public class GeetestConfig {
 
 	// TODO: replace the these two string with your own captcha's id/key,the id/key below is just for demo
-	private static final String captcha_id = "b46d1900d0a894591916ea94ea91bd2c";
-	private static final String private_key = "36fc3fe98530eea08dfc6ce76e3d24c4";
+	private static final String captcha_id = "157e7df54d8deb46238cef3c5848a2bf";
+	private static final String private_key = "f48a9f88c30f4f01696d96ea0d220f98";
 
+
+//    private static final String captcha_id = "b46d1900d0a894591916ea94ea91bd2c";
+//    private static final String private_key = "36fc3fe98530eea08dfc6ce76e3d24c4";
 	public static final String getCaptcha_id() {
 		return captcha_id;
 	}
