+++ src/main/webapp/META-INF/MANIFEST.MF
