+++ src/main/webapp/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>index</title>
+</head>
+<body>
+<a href="/login.jsp">JavaEE站点安装Demo页面1</a><br>
+<p/>
+<a href="/msg_login.jsp">JavaEE站点安装Demo页面2</a>
+</body>
+</html>
